remove!: url parsing error MissingResourceLocation
The `MissingResourceLocation` error variant was previously returned if
the input URL had either scheme ssh (no matter if URL or SCP format) or
git and the URL contained no or an empty path section. In the future we
will instead return the `MissingRepositoryPath` error in such a case.

The only benefit of such a seperation is that it allows the caller to
infer the kind of URL we tried to parse. A future commit will add a new
field to the `MissingRepositoryPath` error variant which will clearly
communicate this information and can therefore be used instead for this
purpose.