remove!: url parsing error NotALocalFile
The `NotALocalFile` error variant was previously returned if the input
url contained a colon but the slice after that character contained no
slash character (`/` or `\`). This behavior is wrong.

SCP like URLs may not need a slash character to specify the repositories
location. Similarly, a local directory that contains a colon in its name
is a valid repository path as well.

The new implementation correctly parses such URLs.