github: rename `jj-docs-bot` to `jj-docs[bot]`
Perhaps this will stop Github from showing jj-docs-bot as a very active
contributor in https://github.com/martinvonz/jj/pulse. This would
probably fair, even though jj-docs-bot tries its best to be a good and
helpful bot.

Regardless, this seems to be the standard on Github which has
`dependabot[bot]` and `github-actions[bot]`.