Fix absolute windows file urls with extra slash
For more context see d6f90beac37866f992a1714d38e5b320eea6f1bb.

We however no longer implement the backslash to forward slash
conversion. There is no good reason for why Git does that conversion
and we do not need it to stay compatible (Git can handle both versions
just fine). Additionally, the logic for the conversion was very involved
and to implement it proper we would need to detect the shell we are
executing in (Git performs the conversion only in bash, not in cmd/ps).